Update .gitignore with specific files and directories
Updated the `.gitignore` file to ignore certain files and directories. These include files in the `.vs` directory, files with extensions `.pyproj`, `.sln`, `.suo`, and `.user`, all files in the `_output` directory, the `custom_config.ini` file in the `config` directory, and the `AutoDrive_config.xml` file in the `_input/map_name` directory of the `AutoDriveTranslationTool` directory. This is to prevent tracking of user-specific settings, build artifacts, or other non-shareable data.