Refactor config setup to centralize "config_name" attribute
This commit refactors the configuration setup of the application. The "config_name" attribute has been removed from individual configuration items and is now passed as a default value in the `ConfigKeyList.add_ConfigKey` and `ConfigHandler.add_variable` methods. This change simplifies the code and improves manageability by defining the configuration name in a single location.