`Added sleep module and updated GUI tasks`
` `

`Added the 'from time import sleep' module to auto_drive_translation_tool.py to pause program execution. Commented out the creation of the LanguagesFrame object and the addition of the 'tab_languages' tab to the tabs dictionary. Added lines to update the GUI window and the translation_frame GUI instance. Also added a line to pause the program for 1 second, potentially allowing the GUI to update before continuing execution.`